requirements.txt
Django==5.0.7
django-ninja==1.2.2
psycopg2==2.9.9
django-extensions==3.2.3